Move it to a single file and finish expand/collapse all
